refactor: steps 처리
